pyproject.toml
[project]
name = "resume-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.4",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "pydantic>=2.6",
    "openai>=1.13",
    "pdfplumber>=0.10",
    "python-docx>=1.1",
    "playwright>=1.42",
    "jinja2>=3.1",
    "python-dotenv==1.2.1",
    "loguru==0.7.3",
    "watchfiles==1.1.1",
    "asyncpg==0.30.0",
    "fastapi==0.121.0",
    "uvicorn==0.38.0",
    "psycopg2-binary==2.9.11",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["bot", "core", "scripts", "backend"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
